requirements.txt
# Core Framework
fastapi>=0.128.0
uvicorn>=0.40.0
pydantic>=2.0.0
python-multipart>=0.0.6

# LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# LLM Providers (Free/Cheap options)
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
openai>=1.0.0

# Frontend/UI
gradio>=5.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.28.0
loguru>=0.7.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
